Some more config, and don't auto connect.
Most of these changes came out of me trying to debug WTF was going on
when the bot wouldn't connect after a netsplit.  Waiting a number of
hours and trying again later seemed to be the "fix", so while these
didn't actually solve the problem, they do seem to make things a bit
more specific and tidy.